Refactor ColumnCollection to inherit from java.util.BitSet, fixing previous refactoring of unionCount() (semantic error) - numberOfColumns is now int (was long) - unionCount() used and to calculate the cardinality, this calculates the cardinality of the intersection => changed to or

+++ dfd/dfdAlgorithm/src/fdiscovery/columns/Path.java
@@ -7,7 +7,7 @@ public class Path extends ColumnCollection {
 
 	private static final long serialVersionUID = -6451347203736964695L;
 
-	public Path(long numberOfColumns) {
+	public Path(int numberOfColumns) {
 		super(numberOfColumns);
 	}
